feat: layout synchronization over TCP
- Server pushes full layout JSON to all connected clients on edit save
- New clients receive the current layout immediately on connection
- Protocol upgraded: payload length uint16→uint32 (header 8→10 bytes)
- New MsgType::LayoutSync (0x0020) with compact JSON payload
- PosServer: setCurrentLayout() caches and broadcasts to all sessions
- PosClient: layoutSyncReceived signal on incoming sync
- MainWindow: layoutChanged signal + applyLayoutFromNetwork()
- Host wires layout changes → server → remote clients
- Client applies received layout automatically
- Each client remains an independent terminal

+++ src/common/Protocol.cpp
@@ -33,7 +33,7 @@ bool deserializeHeader(const QByteArray &data, Header &hdr)
 
     std::uint32_t magic{};
     std::uint16_t type{};
-    std::uint16_t payloadLen{};
+    std::uint32_t payloadLen{};
 
     in >> magic >> type >> payloadLen;
 
@@ -50,7 +50,7 @@ QByteArray buildMessage(MsgType type, const QByteArray &payload)
 {
     Header hdr;
     hdr.type       = type;
-    hdr.payloadLen = static_cast<std::uint16_t>(payload.size());
+    hdr.payloadLen = static_cast<std::uint32_t>(payload.size());
 
     QByteArray msg = serializeHeader(hdr);
     if (!payload.isEmpty())
